transport/tls-session: extract from tlscontext
