+++ lib/transport/Makefile.am
@@ -17,7 +17,8 @@ transportinclude_HEADERS = \
 	lib/transport/transport-factory-tls.h \
 	lib/transport/transport-factory-socket.h \
 	lib/transport/tls-context.h \
-	lib/transport/tls-verifier.h
+	lib/transport/tls-verifier.h \
+	lib/transport/tls-session.h
 
 transport_sources = \
 	lib/transport/logtransport.c	\
@@ -32,7 +33,8 @@ transport_sources = \
 	lib/transport/transport-factory-tls.c \
 	lib/transport/transport-factory-socket.c \
 	lib/transport/tls-context.c \
-	lib/transport/tls-verifier.c
+	lib/transport/tls-verifier.c \
+	lib/transport/tls-session.c
 
 transport_crypto_sources = \
 	lib/transport/transport-tls.c
